update to separate routers for fastapi
